Common: Refactor get_image to common functions
